feat: do not emit "error" events anymore
Previously, the "error" events from the two Redis clients were
forwarded by each instance of the adapter, which made it quite hard to
handle the errors with a lot of namespaces (as there is one instance of
adapter per namespace).

```js
io.of("/my-namespace").adapter.on("error", () => {
  // ...
});
```

The adapter instance will no longer emit "error" events, and will print
a warning if there is no other error handler registered (for backward
compatibility).

The error handling must be done directly on the Redis client:

```js
redisClient.on("error", (err) => {
  // something went wrong, maybe the connection to the server was lost
});
```

Related:

- https://github.com/socketio/socket.io-redis-adapter/issues/412
- https://github.com/socketio/socket.io-redis-adapter/issues/425